requirements/base.txt
bpython==0.12
Django==1.6.1
django-allauth==0.15.0
django-braces==1.2.2
django-filter==0.7
django-model-utils==1.5.0
django-stripe-payments==1.3
django-suit==0.2.5
django-tastypie==0.11.0
djangorestframework==2.3.12
logutils==0.3.3
Markdown==2.3.1
MySQL-python==1.2.5
python-openid==2.2.5
requests==2.2.0
requests-oauthlib==0.4.0
South==0.8.2